previous_pipeline/pyproject.toml
[project]
requires-python = ">= 3.10"
name = "SRTCC"
version = "1.0.1"
readme = "README.md"
license = { text = "pyproject-template © 2024 by RSLab@DISI, UniTrento (Italy) is licensed under CC BY 4.0 " }
description = "SRT project Climate Change"
authors = [{ name = "Luca Zanolo", email = "luca.zanolo@studenti.unitn.it" }]
maintainers = [{ name = "Luca Zanolo", email = "luca.zanolo@studenti.unitn.it" }]
dependencies = [
    "folium == 0.17.0",
    "geopandas == 1.0.1",
    "h5py == 3.11.0",
    "ipykernel == 6.29.5",
    "ipywidgets == 8.1.3",
    "joblib == 1.4.2",
    "matplotlib == 3.7.0",
    "mapclassify == 2.8.0",
    "numpy == 1.24.0",
    "openpyxl == 3.1.5",
    "pandas == 2.2.2",
    "pyogrio == 0.9.0",
    "PyYAML == 6.0.1",
    "rasterio == 1.3.10",
    "scikit-image == 0.24.0",
    "scikit-learn == 1.5.1",
    "scipy == 1.14.0",
    "tqdm == 4.66.5",
    "wheel == 0.44.0",
    "setuptools == 72.1.0",
    "tabulate == 0.9.0",
    "GDAL",
]
